requirements.txt
azure-functions
azure-ai-textanalytics
python-dotenv
